refactor: feature folders, shared components, alias paths; update imports
