+++ src/features/plant/Plant.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import Sidebar from "../Sidebar";
+import Sidebar from "../../shared/components/Sidebar/Sidebar";
 import { styled } from '@mui/material/styles';
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
@@ -16,8 +16,13 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import TablePagination from '@mui/material/TablePagination';
+import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
+import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
+import Collapse from '@mui/material/Collapse';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
 import Paper from '@mui/material/Paper';
-import { db } from "../../firebase/firebase"
+import { db } from "../../shared/services/firebase/firebase"
 import { collection, addDoc, getDocs,deleteDoc, doc,onSnapshot, updateDoc } from "firebase/firestore";
 
 
@@ -40,11 +45,12 @@ const DrawerHeader = styled('div')(({ theme }) => ({
   ...theme.mixins.toolbar,
 }));
 
-const Sku = () => {
+const Plant = () => {
+  const [open, setOpen] = useState(false);
   const [form, setForm] = useState({})
   const [data, setData] = useState([])
   const [editId, setEditId] = useState(null);
-  const sku = collection(db, "sku")
+  const plant = collection(db, "plant")
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(10);
 
@@ -65,7 +71,7 @@ const Sku = () => {
   },[])
 
   const loadRealtime =  () => {
-    const unsubscribe = onSnapshot(sku, (snapshot) => {
+    const unsubscribe = onSnapshot(plant, (snapshot) => {
       const newData = snapshot.docs.map((doc) => ({
         id: doc.id,
         ...doc.data(),
@@ -79,7 +85,7 @@ const Sku = () => {
 
   const handleAddData = async () => {
     if (editId) {
-      const docRef = doc(db, "sku", editId);
+      const docRef = doc(db, "plant", editId);
       await updateDoc(docRef, form)
         .then(() => {
           setEditId(null);
@@ -87,7 +93,7 @@ const Sku = () => {
         })
         .catch((err) => console.log(err));
     } else {
-      await addDoc(sku, form)
+      await addDoc(plant, form)
         .then((res) => {
           setForm({});
         })
@@ -104,7 +110,7 @@ const Sku = () => {
 
   const handleDelete = async (id) => {
     try {
-      await deleteDoc(doc(sku, id));
+      await deleteDoc(doc(plant, id));
     } catch (err) {
       console.log(err);
     }
@@ -116,71 +122,70 @@ const Sku = () => {
   };
   return (
     <>
-      <Box sx={{ display: 'flex' }}>
+      <Box style={{fontFamily: 'Kanit, sans-serif'}} sx={{ display: 'flex' }}>
           <Sidebar />
           <Box component="main" sx={{ flexGrow: 1, p: 3 }}>
           <DrawerHeader />
-          <h1>SKU</h1>
+          <h1>Plant</h1>
           
                 <Box
                   component="form"
                   sx={{
-                    '& > :not(style)': { m: 2, width: '30ch' },
+                    '& > :not(style)': { m: 1, width: '25ch' },
                   }}
                   noValidate
                   autoComplete="off"
                 >
                   <TextField 
                   onChange={(e) => handleChange(e)}
-                  value={form.skuName || ''}
-                  name="skuName" 
+                  value={form.plantName || ''}
+                  name="plantName" 
                   id="outlined-basic" 
-                  label="SKU" 
-                  variant="outlined" />
-
-                  <TextField 
-                  onChange={(e) => handleChange(e)}
-                  value={form.size || ''}
-                  name="size" 
-                  id="outlined-basic" 
-                  label="Size (L/Bottle)" 
+                  label="Plant" 
                   variant="outlined" />
                 </Box>
-                
             
-            <FormControl sx={{ m: 2, width: 400 }}>
+            <FormControl sx={{ m: 1, width: 400 }}>
                 <Stack direction="row" spacing={10}>
               
-                  <Button size="large" variant="contained" onClick={handleAddData} endIcon={<DoneAll />}>
-                    Submit
+                  <Button style={{fontFamily: 'Kanit, sans-serif'}} size="large" variant="contained" onClick={handleAddData} endIcon={<DoneAll />}>
+                    ตกลง
                   </Button>
               </Stack>
             </FormControl>
             <FormControl sx={{ m: 1, width: "100%" }}>
             <TableContainer component={Paper}>
-              <Table sx={{ minWidth: 650 }} aria-label="simple table">
+              <Table sx={{ minWidth: 650 }} aria-label="collapsible table">
                 <TableHead>
                   <TableRow>
-                    <StyledTableCell width={"50%"}><b>SKUs</b></StyledTableCell>
-                    <StyledTableCell width={"50%"}><b>Size (L/Bottle)</b></StyledTableCell>
-                    <StyledTableCell align="center">Edit</StyledTableCell>
-                    <StyledTableCell align="center">Delete</StyledTableCell>
+                    <StyledTableCell />
+                    <StyledTableCell style={{fontFamily: 'Kanit, sans-serif'}} width={"100%"}><b>Plant</b></StyledTableCell>
+                    <StyledTableCell style={{fontFamily: 'Kanit, sans-serif'}} align="center">Edit</StyledTableCell>
+                    <StyledTableCell style={{fontFamily: 'Kanit, sans-serif'}} align="center">Delete</StyledTableCell>
                   </TableRow>
                 </TableHead>
                 <TableBody>
+                
                   {data
                   .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                   .map((row) => (
+                    <React.Fragment>
                     <TableRow
                       key={row.id}
                       tabIndex={-1}
                       sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
                     >
+                        <TableCell>
+                            <IconButton
+                                aria-label="expand row"
+                                size="small"
+                                onClick={() => setOpen(!open)}
+                            >
+                                {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
+                            </IconButton>
+                        </TableCell>
                       <TableCell component="th" scope="row">
-                        {row.skuName}
-                      </TableCell>
-                      <TableCell component="th" scope="row">
-                        {row.size}
+                        <span style={{color: "blue", fontWeight: "bold"}}> {row.plantName} </span>
                       </TableCell>
                       <TableCell align="center">
                         <Button variant="contained" color="primary" onClick={() => handleEdit(row)}>
@@ -193,7 +198,37 @@ const Sku = () => {
                         </Button>
                       </TableCell>
                     </TableRow>
+                     <TableRow>
+                     <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
+                        <Collapse in={open} timeout="auto" unmountOnExit>
+                         <Box sx={{ margin: 1 }}>
+                         <Typography variant="h6" gutterBottom component="div">
+                            ไลน์ผลิต
+                         </Typography>
+                         <Table size="small" aria-label="purchases">
+                             <TableHead>
+                                 <TableRow>
+                                     <TableCell></TableCell>
+                                 </TableRow>
+                             </TableHead>
+                             <TableBody>
+                                 {row.lines.map((line) => (
+                                     <TableRow key={line}>
+                                         <TableCell component="th" scope="row">
+                                             {line}
+                                         </TableCell>
+                                     </TableRow>
+                                 ))}
+ 
+                             </TableBody>
+                         </Table>
+                         </Box>
+                         </Collapse>
+                     </TableCell>
+                   </TableRow>
+                   </React.Fragment>
                   ))}
+                 
                 </TableBody>
               </Table>
             </TableContainer>
@@ -214,4 +249,4 @@ const Sku = () => {
   );
 };
 
-export default Sku;
+export default Plant;
