+++ src/shared/components/Table/ProductionHrCell.jsx
@@ -1,32 +1,29 @@
 import React, { useState, useEffect } from 'react';
 import { TextField, TableCell } from '@mui/material';
 import { doc, updateDoc } from "firebase/firestore";
-import { db } from "../../firebase/firebase"
+import { db } from "../../services/firebase/firebase"
 import Tooltip from '@mui/material/Tooltip';
 
-const ForecastCell = ({ documentId, plantIndex, lineIndex, line, detail, rowSpan, detailIndex, onUpdate }) => {
+const ProductionHrCell = ({ documentId, plantIndex, lineIndex, line, detail, detailIndex, onUpdateProductionHr }) => {
   const [isEditing, setIsEditing] = useState(false);
-  const [value, setValue] = useState(detail.lineForecast)
-  // const [value, setValue] = useState(
-  //   line.groupProducts.includes("Beer") ? line.forecast : detail.lineForecast
-  // );
-
-  useEffect(() => {
-    setValue(detail.lineForecast);
-  }, [detail.lineForecast]);
+  const [value, setValue] = useState(detail.productionHrStd);
 
   const handleEdit = () => {
     setIsEditing(true);
   };
 
+  useEffect(() => {
+    setValue(detail.productionHrStd);
+  }, [detail.productionHrStd]);
+
   const handleBlur = async () => {
     setIsEditing(false);
-    await onUpdate(documentId, plantIndex, lineIndex, detailIndex, line, value);
+    await onUpdateProductionHr(documentId, plantIndex, lineIndex, detailIndex, line, value);
   };
 
   const handleKeyPress = async (e) => {
     if (e.key === 'Enter') {
-        await onUpdate(documentId, plantIndex, lineIndex, detailIndex, line, value);
+        await onUpdateProductionHr(documentId, plantIndex, lineIndex, detailIndex, line, value);
       setIsEditing(false);
     }
   };
@@ -37,8 +34,7 @@ const ForecastCell = ({ documentId, plantIndex, lineIndex, line, detail, rowSpan
 
   return (
     <>
-    
-        <TableCell align="center" style={{ color: "#8A2BE2", fontWeight: "bold"}}>
+        <TableCell align="center" style={{color: "#10AC85", fontWeight: "bold"}}>
            {isEditing ? (
                 <TextField
                 value={value}
@@ -50,14 +46,14 @@ const ForecastCell = ({ documentId, plantIndex, lineIndex, line, detail, rowSpan
                 size="small"
                 />
             ) : (
-                <span onClick={handleEdit} rowSpan={1} style={{ cursor: 'pointer' }}>
-                   <Tooltip title="คลิ๊กเพื่ออัพเดท">{value}</Tooltip>
+                <span onClick={handleEdit} style={{ cursor: 'pointer' }}>
+                    <Tooltip title="คลิ๊กเพื่ออัพเดท">{value}</Tooltip>
                 </span>
             )}
          </TableCell>
-        </>
-   
+
+    </>
    );
 };
 
-export default ForecastCell
+export default ProductionHrCell
