+++ src/shared/components/Table/TablePlantCell.jsx
