+++ src/shared/services/firebase/auth.js
